checkpoint - tentando arrumar o login e logout

+++ front-end/src/app/auth/login/login.component.ts
@@ -1,13 +1,10 @@
-import Swal from 'sweetalert2';
-import { DashboardService } from './../../dashboard/services/dashboard.service';
-import { Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormControl, FormsModule } from '@angular/forms';
-import { first } from 'rxjs/operators'
-import { AppComponent } from 'src/app/app.component';
+import { FormGroup } from '@angular/forms';
+import { Router } from '@angular/router';
 import { CookieService } from 'ngx-cookie-service';
-import { JwtHelperService } from '@auth0/angular-jwt';
+import { AuthService } from '../services/auth.service';
+import { DashboardService } from './../../dashboard/services/dashboard.service';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -15,7 +12,7 @@ import { JwtHelperService } from '@auth0/angular-jwt';
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.css']
 })
-export class LoginComponent {
+export class LoginComponent implements OnInit {
 
   username: string;
   password: string;
@@ -33,24 +30,35 @@ export class LoginComponent {
   }
 
   onSubmit() {
-      this.authService.login(this.username, this.password).subscribe(
-        response => {
-          const token = response.jwt;
-          localStorage.setItem('token', token);
+    this.authService.login(this.username, this.password).subscribe(
+      response => {
+        //exibe um alerta de sucesso e redireciona para a página de dashboard
+        Swal.fire({
+          icon: 'success',
+          title: 'Login realizado com sucesso',
+          timer: 1000,
+          showConfirmButton: false
+        }).then(() => {
           this.router.navigate(['/dashboard']);
           console.log(this.authService.getUsername());
-
-
-        },
+          const token = response.jwt;
+          localStorage.setItem('token', token);
+        });
+        // const navigationExtras: NavigationExtras = {
+        //   queryParams: { id },
+        // };
+        // this.router.navigate(['/dashboard']);
+        // console.log(this.authService.getUsername());
+      },
       error => {
         // exibe um alerta de erro
-        // Swal.fire({
-        //   icon: 'error',
-        //   title: 'Erro ao fazer login',
-        //   text: 'Usuário ou senha incorretos',
-        //   confirmButtonColor: '#3085d6',
-        //   confirmButtonText: 'Ok'
-        // });
+        Swal.fire({
+          icon: 'error',
+          title: 'Erro ao fazer login',
+          text: 'Usuário ou senha incorretos',
+          confirmButtonColor: '#EDA900',
+          confirmButtonText: 'Ok'
+        });
       }
     );
   }
